+++ CHANGES.md
@@ -21,6 +21,7 @@ Release Notes.
 * Add the `@SuperDataset` annotation for BrowserErrorLog.
 * Add the thread pool to the Kafka fetcher to increase the performance.
 * Add `contain` and `not contain` OPS in OAL.
+* Add Envoy ALS analyzer based on metadata exchange.
 * Support keeping collecting the slowly segments in the sampling mechanism.
 * Support choose files to active the meter analyzer.
 * Improve Kubernetes service registry for ALS analysis.